:truck: refactor : 테스트코드 파일 패키지 변경
